Split _core.py into _parsing, _preamble, _comments_resolve, _literalize
Drops leading underscore from cross-module symbols to satisfy
reportPrivateUsage (_get_yaml, _parse_input, _compute_preamble,
_ResolvedComments, _resolve_{yaml,toml}_comments).
